# api/openapi.yaml
...
openapi: 3.1.0
info:
  title: IronClaw Control-Plane API
  version: 0.1.0
  summary: Gateway change-approval, audit, and registry administration for a self-hosted IronClaw control-plane.
  description: |
    The control-plane exposes three surfaces:

    - **Gateway** (`/v1/changes`, `/v1/audit`) — submit capability changes, list
      what is pending, record human approve/reject decisions, and read the
      append-only audit log. Every control-plane mutation (persona, enabled
      tools, packages, wiring, permissions, mounts, create_agent) flows through
      here; there is no bypass path.
    - **Registry** (`/v1/registry/...`) — administer agent groups, messaging
      groups, wirings, users, roles, members, destinations, and read sessions.
      These endpoints return `503` until a registry backend is attached.
    - **Operational probes** (`/healthz`, `/readyz`, `/metrics`).

    All `POST`/`PUT` bodies are JSON. The Go JSON decoder is case-insensitive on
    keys; the property names below are the canonical forms the server **emits**.
  license:
    name: MIT
    url: https://github.com/nivardsec/ironclaw/blob/main/LICENSE
  contact:
    name: IronClaw
    url: https://github.com/nivardsec/ironclaw
...
servers:
  - url: http://127.0.0.1:8787
    description: Local development (`--dev`, loopback bind).
  - url: http://{tailnetHost}:{port}
    description: Production — bound to the control-plane's Tailscale interface address (no public port).
    variables:
...
      port:
...

...
security:
  - bearerAuth: []
...
paths:
...
  /v1/registry/mcp-servers:
    get:
      tags: [mcp]
      summary: List configured MCP servers (secrets masked).
      description: Returns every server in the catalog with secret-bearing env/header values masked. Requires MCP enabled (--mcp-catalog), else 503.
      operationId: listMCPServers
      responses:
        "200":
          description: The configured servers.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MCPServer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/MCPUnavailable"
  /v1/registry/mcp-servers/{name}:
    put:
      tags: [mcp]
      summary: Create or update an MCP server.
      description: The path `name` is authoritative. A secret value submitted as the mask placeholder (`••••`) is treated as unchanged and the stored value is preserved. Configuring a server grants no agent anything — a gateway-approved `mcp_access` change does that.
      operationId: putMCPServer
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MCPServer"
      responses:
        "200":
          description: The stored server (secrets masked).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MCPServer"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/MCPUnavailable"
    get:
      tags: [mcp]
      summary: Fetch one MCP server (secrets masked).
      operationId: getMCPServer
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The server.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MCPServer"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "503":
          $ref: "#/components/responses/MCPUnavailable"
    delete:
      tags: [mcp]
      summary: Remove an MCP server from the catalog.
      description: Does not revoke already-approved grants (a separate gateway change); the broker simply can no longer reach a removed server.
      operationId: deleteMCPServer
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "204":
          description: Removed.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/MCPUnavailable"
  /v1/registry/mcp-servers/{name}/probe:
    post:
      tags: [mcp]
      summary: Connect to a server and discover its tools.
      description: An operator action on infrastructure (it does not consult grants). Spawns/dials the server, performs the MCP handshake, and returns its declared tools.
      operationId: probeMCPServer
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: The server's declared tools.
          content:
            application/json:
              schema:
                type: object
                properties:
                  tools:
                    type: array
                    items:
                      $ref: "#/components/schemas/MCPTool"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "502":
          description: The server could not be reached or did not complete the handshake.
        "503":
          $ref: "#/components/responses/MCPUnavailable"
  /v1/ui/mcp-servers:
    get:
      tags: [mcp]
      summary: Console read-model — servers plus the agent groups granted each.
      operationId: uiMCPServers
      responses:
        "200":
          description: Servers (secrets masked) with their grants.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    server:
                      $ref: "#/components/schemas/MCPServer"
                    grants:
                      type: array
                      items:
                        type: object
                        properties:
                          agentGroupId: { type: string }
                          agentGroupName: { type: string }
                          tools:
                            type: array
                            items: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/MCPUnavailable"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token, supplied to the server via `IRONCLAW_API_TOKEN`
        (or `--token` for `ironctl`). It is **all-or-nothing** — there are no
        per-endpoint scopes; a valid token authorizes every non-probe endpoint.
        The token is compared in constant time. `/healthz` and `/readyz` are
        always exempt. When the server is started without a token the check is
        disabled and the mesh (Tailscale) boundary is the sole control.
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token (only when the server is started with a token).
      content:
        text/plain:
          schema:
            type: string
            example: unauthorized
    BadRequest:
      description: Malformed JSON body or invalid field values.
      content:
        text/plain:
          schema:
            type: string
    NotFound:
      description: Resource not found.
      content:
        text/plain:
          schema:
            type: string
...
    MCPUnavailable:
      description: MCP is not enabled on this control-plane (start the daemon with --mcp-catalog).
      content:
        text/plain:
          schema:
            type: string
            example: MCP is not enabled on this control-plane (set --mcp-catalog)
  schemas:
...
    MCPServer:
      type: object
      description: An operator-configured MCP server. Local (stdio subprocess) or remote (streamable HTTP). Secret values in env/headers should be ${ENV} references (resolved host-side); raw values are masked on read.
      required: [name, transport]
      properties:
        name:
          type: string
          description: Unique lowercase label.
        displayName: { type: string }
        description: { type: string }
        transport:
          type: string
          enum: [stdio, http]
          description: stdio = a local subprocess the host runs (isolated); http = a remote endpoint the host dials over TLS.
        command:
          type: string
          description: (stdio) the server executable.
        args:
          type: array
          items: { type: string }
          description: (stdio) arguments.
        env:
          type: object
          additionalProperties: { type: string }
          description: (stdio) extra environment; values may be ${ENV} references.
        image:
          type: string
          description: (stdio) container image for isolation; empty uses the daemon's --mcp-image.
        url:
          type: string
          description: (http) endpoint URL; https required for a non-loopback host.
        headers:
          type: object
          additionalProperties: { type: string }
          description: (http) request headers; values may be ${ENV} references.
    MCPTool:
      type: object
      description: One tool a server declares (the tools/list element).
      properties:
        name: { type: string }
        description: { type: string }
        inputSchema:
          type: object
          description: JSON Schema for the tool's arguments, passed through from the server.
...